requirements.txt
streamlit>=1.28
google-genai>=0.3
Pillow>=10.0
